+++ VRCOSC.Game/Modules/ChatBox/ChatBoxModule.cs
@@ -9,7 +9,7 @@ public abstract class ChatBoxModule : Module
 {
     protected void CreateVariable(Enum lookup, string name, string format) => ChatBoxManager.RegisterVariable(SerialisedName, lookup.ToLookup(), name, format);
     protected void CreateState(Enum lookup, string name, string defaultFormat) => ChatBoxManager.RegisterState(SerialisedName, lookup.ToLookup(), name, defaultFormat);
-    protected void CreateEvent(Enum lookup, string name, string defaultFormat, int defaultLength) => ChatBoxManager.RegisterEvent(SerialisedName, lookup.ToLookup(), name, defaultFormat, defaultLength);
+    protected void CreateEvent(Enum lookup, string name, string defaultFormat, float defaultLength) => ChatBoxManager.RegisterEvent(SerialisedName, lookup.ToLookup(), name, defaultFormat, defaultLength);
 
     protected void SetVariableValue(Enum lookup, string? value, string suffix = "") => ChatBoxManager.SetVariable(SerialisedName, lookup.ToLookup(), value, suffix);
     protected string GetVariableFormat(Enum lookup) => ChatBoxManager.VariableMetadata[SerialisedName][lookup.ToLookup()].DisplayableFormat;
